Move utils to utils directory
