+++ app/components/NavigationHeader.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import { cn } from "@/lib/utils"
 </script>
 
 <template>
