+++ app/components/ui/button/Button.vue
@@ -2,7 +2,7 @@
 import type { HTMLAttributes } from "vue"
 import { Primitive, type PrimitiveProps } from "radix-vue"
 import { type ButtonVariants, buttonVariants } from "."
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/utils"
 
 interface Props extends PrimitiveProps {
   variant?: ButtonVariants["variant"]
