+++ app/components/ui/dropdown-menu/DropdownMenuLabel.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { DropdownMenuLabel, type DropdownMenuLabelProps, useForwardProps } from "radix-vue"
 import { computed, type HTMLAttributes } from "vue"
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/utils"
 
 const props = defineProps<DropdownMenuLabelProps & { class?: HTMLAttributes["class"], inset?: boolean }>()
