+++ app/components/ui/dropdown-menu/DropdownMenuShortcut.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import type { HTMLAttributes } from "vue"
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/utils"
 
 const props = defineProps<{
   class?: HTMLAttributes["class"]
