+++ components.json
@@ -13,6 +13,6 @@
   "framework": "nuxt",
   "aliases": {
     "components": "@/components",
-    "utils": "@/lib/utils"
+    "utils": "@/utils/utils"
   }
 }
